requirements.txt
Django>=4.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.7
